+++ src/core.js
@@ -9,6 +9,7 @@
         //noinspection JSValidateTypes
         var supportPageOffset = window.pageXOffset !== undefined,
             isCSS1Compatible = ((document.compatMode || "") === "CSS1Compat"),
+            agent = window.navigator.userAgent,
             noop = function(){},
             okanjo = {
 
@@ -278,27 +279,31 @@
 
 
                     /**
-                     * Cross browser way to get the height of an element
-                     * @param {HTMLElement} el – The DOM element to get the height of
+                     * Gets the height and width of the given element
+                     * @param {HTMLElement|Node} el – The DOM element to get the size of
                      * @param {boolean} [includeMargin] – Whether to include the margins of the element in the size
-                     * @returns {number} – Height of the element
+                     * @returns {{height: number, width: number}}
                      */
-                    getOuterHeight: function(el, includeMargin) {
-                        if (includeMargin) {
-                            var height = el.offsetHeight;
-                            var style = el.currentStyle || getComputedStyle(el);
+                    getElementSize: function(el, includeMargin) {
 
-                            height += parseInt(style.marginTop) + parseInt(style.marginBottom);
-                            return height;
-                        } else {
-                            return el.offsetHeight;
+                        var size = {
+                            height: el.offsetHeight,
+                            width : el.offsetWidth
+                        }, style;
+
+                        if (includeMargin) {
+                            style = el.currentStyle || getComputedStyle(el);
+                            size.height += parseInt(style.marginTop) + parseInt(style.marginBottom);
+                            size.width += parseInt(style.marginLeft) + parseInt(style.marginRight);
                         }
+
+                        return size;
                     },
 
 
                     /**
                      * Splits the text in the element to fit within the visible height of its container, and separates with an ellipses
-                     * @param {HTMLElement} element – The DOM element containing the text to fit
+                     * @param {HTMLElement|Node} element – The DOM element containing the text to fit
                      * @param {HTMLElement} [container] – Optional container to compute fit on. Defaults to the element's parent
                      */
                     ellipsify: function(element, container) {
@@ -306,7 +311,7 @@
                         // It's a sad day when you have to resort to JS because CSS kludges are too hacky to work down to IE8, programmatically
                         //noinspection JSValidateTypes
                         var parent = container || element.parentNode,
-                            targetHeight = okanjo.util.getOuterHeight(parent),
+                            targetHeight = okanjo.util.getElementSize(parent).height,
                             useTextContent = element.textContent !== undefined,
                             text = useTextContent ? element.textContent : element.innerText,
                             replacedText = "",
@@ -318,7 +323,7 @@
 
                         // Trim off characters until we can fit the text and ellipses
                         // If the text already fits, this loop is ignored
-                        while (okanjo.util.getOuterHeight(element) > targetHeight && text.length > 0 && (safety-- > 0)) {
+                        while (okanjo.util.getElementSize(element).height > targetHeight && text.length > 0 && (safety-- > 0)) {
                             text = useTextContent ? element.textContent : element.innerText;
 
                             text = text.replace(/[\s\S](?:\.\.\.)?$/, replacer);
@@ -372,6 +377,41 @@
                             classDetects.push('lt-ie7');
                         }
                         return classDetects;
+                    },
+
+
+                    /**
+                     * Checks if we're executing this code inside a frame or not
+                     * @returns {boolean}
+                     */
+                    isFramed: function() {
+                        return window.top !== window.self;
+                    },
+
+                    /**
+                     * Checks if the current user agent identifies as an iOS device
+                     * @returns {boolean}
+                     */
+                    isiOS: function() {
+                        return /(iPhone|iPad|iPod)/i.test(agent);
+                    },
+
+
+                    /**
+                     * Checks if the current user agent identifies as Android device
+                     * @returns {boolean}
+                     */
+                    isAndroid: function() {
+                        return /Android/.test(agent);
+                    },
+
+
+                    /**
+                     * Checks if the current user agent identifies as a mobile device
+                     * @returns {boolean}
+                     */
+                    isMobile: function() {
+                        return okanjo.util.isiOS() || okanjo.util.isAndroid();
                     }
 
                 }